Refactor preferences and add Other group
Renamed Preferences to PDEPreferences and moved the 'Other' preference group logic from General.kt to a new Other.kt file for better separation. Registered the Other group in PDEPreferences to improve organization and maintainability of preference groups.
